requirements.txt
influxdb==5.2.0
kafka-python==1.4.3
PyYAML==3.13
redis==3.0.1